refactor: fcfsissue 모듈 생성
core, coupon 모듈을 의존하는 최상위 application 모듈
