+++ build.gradle.kts
@@ -16,7 +16,9 @@ allprojects {
     repositories {
         mavenCentral()
     }
+}
 
+subprojects {
     apply(plugin = "java-library")
     apply(plugin = "org.springframework.boot")
     apply(plugin = "io.spring.dependency-management")
@@ -24,33 +26,4 @@ allprojects {
     tasks.withType<Test> {
         useJUnitPlatform()
     }
-}
-
-dependencies {
-    implementation(project(":core"))
-    implementation(project(":coupon"))
-
-    implementation("org.springframework.boot:spring-boot-starter-web")
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
-
-    // spring data redis
-    implementation("org.springframework.boot:spring-boot-starter-data-redis")
-
-    // spring data mongodb
-    implementation("org.springframework.boot:spring-boot-starter-data-mongodb")
-
-    // spring validation
-    implementation("org.springframework.boot:spring-boot-starter-validation")
-
-    // spring mail
-    implementation("org.springframework.boot:spring-boot-starter-mail")
-
-    // embedded redis
-    implementation("it.ozimov:embedded-redis:0.7.2")
-
-    // embedded mongodb
-    implementation("de.flapdoodle.embed:de.flapdoodle.embed.mongo.spring30x:4.6.2")
-
-    // rest-assured
-    testImplementation("io.rest-assured:rest-assured:5.3.0")
 }
