+++ core/src/main/java/com/coupop/core/inputvalidation/EmailValidation.java
@@ -1,0 +1,7 @@
+package com.coupop.core.inputvalidation;
+
+public class EmailValidation {
+
+    public static final String REGEX = "^[a-zA-Z0-9+-\\_.]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$";
+    public static final String MESSAGE = "형식에 맞는 이메일을 입력하세요.";
+}
