+++ coupon/src/main/java/com/coupop/coupon/dto/HistoryRequest.java
@@ -1,6 +1,6 @@
 package com.coupop.coupon.dto;
 
-import com.coupop.coupon.support.EmailValidation;
+import com.coupop.core.inputvalidation.EmailValidation;
 import jakarta.validation.constraints.Email;
 
 public record HistoryRequest(@Email(regexp = EmailValidation.REGEX, message = EmailValidation.MESSAGE) String email) {
