+++ fcfs-issue/src/main/java/com/coupop/fcfsissue/dto/IssuanceRequest.java
@@ -1,0 +1,7 @@
+package com.coupop.fcfsissue.dto;
+
+import com.coupop.core.inputvalidation.EmailValidation;
+import jakarta.validation.constraints.Email;
+
+public record IssuanceRequest(@Email(regexp = EmailValidation.REGEX, message = EmailValidation.MESSAGE) String email) {
+}
