+++ fcfs-issue/src/test/java/com/coupop/FcfsCouponApplicationTests.java
@@ -1,0 +1,13 @@
+package com.coupop;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FcfsCouponApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
