+++ settings.gradle.kts
@@ -1,2 +1,2 @@
 rootProject.name = "fcfs-coupon"
-include("core", "coupon")
+include("core", "coupon", "fcfs-issue")
