--- src/main/java/com/coupop/fcfscoupon/fcfsissue/model/FcfsIssueRepository.java
@@ -1,10 +1,0 @@
-package com.coupop.fcfscoupon.fcfsissue.model;
-
-public interface FcfsIssueRepository {
-
-    Long add(final String email);
-
-    Long remove(final String email);
-
-    Long getCount();
-}
