--- src/test/java/com/coupop/fcfscoupon/FcfsCouponApplicationTests.java
@@ -1,13 +1,0 @@
-package com.coupop.fcfscoupon;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FcfsCouponApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
