--- src/test/java/com/coupop/fcfscoupon/testconfig/IntegrationTestConfig.java
@@ -1,43 +1,0 @@
-package com.coupop.fcfscoupon.testconfig;
-
-
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.BDDMockito.given;
-
-import com.coupop.coupon.model.CouponEmailSender;
-import com.coupop.coupon.model.RandomCodeGenerator;
-import com.coupop.fcfscoupon.fcfsissue.model.FcfsIssuePolicy;
-import com.coupop.fcfscoupon.fcfsissue.support.RequestTime;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-
-@SpringBootTest
-public abstract class IntegrationTestConfig {
-
-    protected static final String MOCKED_COUPON_VALUE = "fakevalue";
-
-    @Autowired
-    protected DatabaseSetUp databaseSetUp;
-
-    @MockBean
-    protected RequestTime requestTime;
-
-    @MockBean
-    protected RandomCodeGenerator codeGenerator;
-
-    @MockBean
-    protected CouponEmailSender couponEmailSender;
-
-    @BeforeEach
-    void integrationSetup() {
-        databaseSetUp.clean();
-
-        given(requestTime.getValue())
-                .willReturn(FcfsIssuePolicy.getOpenAt());
-
-        given(codeGenerator.generate(anyLong()))
-                .willReturn(MOCKED_COUPON_VALUE);
-    }
-}
